refactor: NumbersComparator 클래스 생성 후 BaseballGame 클래스에 있는 숫자 비교 메소드 코드 이동
