+++ src/main/java/baseball/Application.java
@@ -1,20 +1,19 @@
 package baseball;
 
 import baseball.domain.BaseballGame;
-import baseball.view.InputView;
 
 import java.util.Scanner;
 
 public class Application {
-    private static final String END = "2";
+
 
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
         // TODO 구현 진행
-        BaseballGame baseballGame = new BaseballGame();
+        BaseballGame baseballGame = new BaseballGame(scanner);
         do {
-            baseballGame.play(scanner);
-        } while (!InputView.inputRestartOrEnd(scanner).equals(END));
+            baseballGame.play();
+        } while (baseballGame.isRestart());
     }
 
 }
