+++ src/main/java/baseball/domain/Computer.java
@@ -9,24 +9,18 @@ public class Computer {
     private final int NUMBERS_LENGTH = 3;
     private final int MIN_NUMBER = 1;
     private final int MAX_NUMBER = 9;
-    private StringBuilder computerNumbers;
+    private StringBuilder stringBuilder;
     private Set<Integer> numbersSet;
 
-    public Computer() {
-        numbersSet = new HashSet<>();
-    }
-
     public String createRandomComputerNumbers() {
-        computerNumbers = new StringBuilder();
+        numbersSet = new HashSet<>();
+        stringBuilder = new StringBuilder();
         while (numbersSet.size() < NUMBERS_LENGTH) {
             numbersSet.add(RandomUtils.nextInt(MIN_NUMBER, MAX_NUMBER));
         }
         for (Integer number : numbersSet) {
-            computerNumbers.append(number);
+            stringBuilder.append(number);
         }
-        numbersSet.clear();
-        return computerNumbers.toString();
+        return stringBuilder.toString();
     }
-
-
 }
