+++ src/main/java/baseball/view/InputView.java
@@ -5,10 +5,8 @@ import utils.InputValidator;
 import java.util.Scanner;
 
 public class InputView {
-
     private static final String INPUT_NUMBER_MESSAGE = "숫자를 입력해주세요 : ";
     private static final String RESTART_OR_END_MESSAGE = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
-
     private static final String INCORRECT_MENU_NUM_ERROR = "1번 또는 2번을 입력해주세요.";
     private static final String RESTART = "1";
     private static final String END = "2";
