+++ src/main/java/baseball/view/OutputView.java
@@ -8,7 +8,6 @@ public class OutputView {
     private static final String NOTHING_MESSAGE = "낫싱";
     private static final String CORRECT_ANSWER_MESSAGE = "3개의 숫자를 모두 맞히셨습니다! 게임 종료";
 
-
     public static void printStrikeBall(int strikeCount, int ballCount) {
         if (strikeCount > ZERO && ballCount == ZERO) {
             System.out.printf(STRIKE_MESSAGE, strikeCount);
